feat: profile-gated guest mode elimination and landing page for feature profile
- Backend (SecurityConfig): inject GalleryProperties and branch permitAll() rules
  by profile (guest access local-only, seed import denyAll() on feature)
- Backend (GalleryProperties): add guestAccessEnabled/seedImportEnabled boolean
  properties (local: true/true, feature: false/false)
- Frontend (environment): add NEXT_PUBLIC_GUEST_MODE and NEXT_PUBLIC_SEED_IMPORT_ENABLED
  flags to compose.yaml (local: true/true) and compose.feature.override.yaml (false/false)
- Frontend (env files): add new flags to .env.local
- Frontend (HomeGate): new client component at root route; branches on profile/auth state
  (local always gallery, feature: logged-out → landing page, logged-in → gallery)
- Frontend (LandingPage): new marketing/dashboard page with hero, features, 'how it works',
  CTAs, and footer (styled consistently with existing dark theme)
- Frontend (GalleryPage): gate seed-import trigger with NEXT_PUBLIC_SEED_IMPORT_ENABLED flag
- Frontend (LoginPage): gate guest hint text with NEXT_PUBLIC_GUEST_MODE flag
- Frontend (page.tsx): swap GalleryPage for HomeGate at root route

Local profile: guest browsing/upload/seed-import unchanged (developer sandbox preserved)
Feature profile: login required for all gallery access; guest mode completely disabled

+++ ImageAPI/src/main/resources/db/migration/V8__add_video_duration.sql
@@ -1,2 +1,2 @@
 -- Add video_duration column to store video length in seconds
-ALTER TABLE images ADD COLUMN video_duration DOUBLE;
+ALTER TABLE images ADD COLUMN video_duration DOUBLE PRECISION;
